Add tests, refactor code
