+++ .eslintrc
@@ -7,6 +7,15 @@
   "parserOptions": {
     "ecmaVersion": 2021
   },
+  "overrides": [
+    {
+      "files": ["tests/**/*"],
+      "plugins": ["jest"],
+      "env": {
+        "jest": true
+      }
+    }
+  ],
   "rules": {
     "arrow-spacing": ["warn", { "before": true, "after": true }],
     "comma-dangle": ["error", "always-multiline"],
